Ignore annotation category or priority filter if not defined
Before this commit, Edirom Online assumed that an annotation would only be shown if it matched both category and priority filters; this was a problem for editions that would not use, e.g., priorities at all. Consequently, the priorities filter would be empty, and the annotation would always return false for that filter.
After this commit, there is a value 'undefined' for visibleCategorie and visibleAnnotations, and the filter corresponding result for an annotation will be set to true if the corresponding visibleCategorie or visiblePriorities is 'undefined',